requirements.txt
python-dotenv~=0.21.1
discord.py
python-dateutil
parsedatetime
typing-extensions
asyncpg
prisma
aiofiles